Refactor of resource and integration connectors
