+++ jdbc-integration-connector/src/main/java/org/odpi/openmetadata/adapters/connectors/integration/jdbc/JdbcIntegrationConnectorProvider.java
@@ -5,7 +5,7 @@ package org.odpi.openmetadata.adapters.connectors.integration.jdbc;
 import org.odpi.openmetadata.frameworks.connectors.ConnectorProviderBase;
 import org.odpi.openmetadata.frameworks.connectors.properties.beans.ConnectorType;
 
-public class JdbcDatabaseConnectorProvider extends ConnectorProviderBase {
+public class JdbcIntegrationConnectorProvider extends ConnectorProviderBase {
 
     static final String  connectorTypeGUID = "49cd6772-1efd-40bb-a1d9-cc9460962ff6";
     static final String  connectorTypeName = "JDBC Database Connector";
@@ -15,8 +15,8 @@ public class JdbcDatabaseConnectorProvider extends ConnectorProviderBase {
      * Constructor used to initialize the ConnectorProviderBase with the Java class name of the specific
      * store implementation.
      */
-    public JdbcDatabaseConnectorProvider(){
-        super.setConnectorClassName(JdbcDatabaseConnector.class.getName());
+    public JdbcIntegrationConnectorProvider(){
+        super.setConnectorClassName(JdbcIntegrationConnector.class.getName());
 
         ConnectorType connectorType = new ConnectorType();
         connectorType.setType(ConnectorType.getConnectorTypeType());
