+++ jdbc-integration-connector/src/main/java/org/odpi/openmetadata/adapters/connectors/integration/jdbc/transfer/TableTransfer.java
@@ -4,8 +4,8 @@ package org.odpi.openmetadata.adapters.connectors.integration.jdbc.transfer;
 
 import org.odpi.openmetadata.accessservices.datamanager.metadataelements.DatabaseTableElement;
 import org.odpi.openmetadata.accessservices.datamanager.properties.DatabaseTableProperties;
+import org.odpi.openmetadata.adapters.connectors.integration.jdbc.transfer.model.JdbcTable;
 import org.odpi.openmetadata.adapters.connectors.integration.jdbc.transfer.requests.Omas;
-import org.odpi.openmetadata.adapters.connectors.resource.jdbc.model.JdbcTable;
 import org.odpi.openmetadata.frameworks.auditlog.AuditLog;
 
 import java.util.List;
